完成医保卡号保存获取等。

+++ SDK_Library/src/main/java/com/wondersgroup/android/jkcs_sdk/ui/leavehospital/view/LeaveHospitalActivity.java
@@ -196,7 +196,7 @@ public class LeaveHospitalActivity extends MvpBaseActivity<LeaveHospitalContract
     @Override
     public void onYiBaoTokenResult(String token) {
         mYiBaoToken = token;
-        if (!TextUtils.isEmpty(mFeeNeedCashTotal) && Long.parseLong(mFeeNeedCashTotal) > 0) {
+        if (!TextUtils.isEmpty(mFeeNeedCashTotal) && Double.parseDouble(mFeeNeedCashTotal) > 0) {
             // 获取支付参数
             mPresenter.getPayParam(mOrgCode);
         } else {
